% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.Dtest}
\alias{plot.Dtest}
\title{Plot method for \code{Dtest} objects}
\usage{
\method{plot}{Dtest}(x, ...)
}
\arguments{
\item{x}{A \code{"Dtest"} object.}

\item{...}{Additional arguments passed to \code{\link{plot_perm_hist}}.}
}
\value{
Invisibly returns the \code{ggplot2} object.
}
\description{
Displays the permutation null distribution histogram with the observed
test statistic annotated.
}
